Tolk v0.6 implementation
LSP internals are fully refactored.
Quite a small amount of original FunC plugin code is left.
A plugin for Tolk is more powerful and fully covers Tolk grammar.
It will be kept up to date with Tolk evolution.

Historically, this extension was developed for several months,
in sync with the development of the Tolk compiler.
But for the first public release, that dev history is not needed.
That's why I just do a single giant commit.